requirements.txt
asgiref==3.11.1
black==26.3.1
click==8.3.1
colorama==0.4.6
cssbeautifier==1.15.4
Django==6.0.2
djangorestframework==3.16.1
djlint==1.36.4
EditorConfig==0.17.1
flake8==7.3.0
isort==8.0.1
jsbeautifier==1.15.4
json5==0.13.0
mccabe==0.7.0
mypy_extensions==1.1.0
mysqlclient==2.2.8
packaging==26.0
pathspec==1.0.4
pillow==12.1.1
platformdirs==4.9.4
pycodestyle==2.14.0
pyflakes==3.4.0
pytokens==0.4.1
PyYAML==6.0.3
regex==2026.2.28
six==1.17.0
sqlparse==0.5.5
tqdm==4.67.3
tzdata==2025.3
